Replace use of uint_fast32_t with size_t.

+++ src/refresh/world.c
@@ -269,7 +269,7 @@ static void GL_MarkLeaves(void)
     byte vis2[VIS_MAX_BYTES];
     mleaf_t *leaf;
     mnode_t *node;
-    uint_fast32_t *src1, *src2;
+    size_t *src1, *src2;
     int cluster1, cluster2, longs;
     vec3_t tmp;
     int i;
@@ -316,8 +316,8 @@ static void GL_MarkLeaves(void)
     if (cluster1 != cluster2) {
         BSP_ClusterVis(bsp, vis2, cluster2, DVIS_PVS);
         longs = VIS_FAST_LONGS(bsp);
-        src1 = (uint_fast32_t *)vis1;
-        src2 = (uint_fast32_t *)vis2;
+        src1 = (size_t *)vis1;
+        src2 = (size_t *)vis2;
         while (longs--) {
             *src1++ |= *src2++;
         }
